requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
faiss-cpu==1.7.4
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
reportlab==4.0.7
numpy==1.24.3
sentence-transformers==2.2.2
huggingface_hub==0.13.4
requests==2.31.0
matplotlib==3.7.2
Pillow==10.0.1
torch==2.1.2
pydantic>=2.4,<3
